% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaGaDecode.R
\name{Gray2Bin}
\alias{Gray2Bin}
\title{Map Gray code to binary.}
\usage{
Gray2Bin(x)
}
\arguments{
\item{x}{Gray code (boolean vector).}
}
\value{
Binary code (boolean vector).
}
\description{
Map Gray code to binary.
}
\details{
Start with the highest order bit, and 
         \code{r[k-i]<- xor(n[k], n[k-1])}.
}
\examples{
 Gray2Bin(c(1, 0, 0, 0))
 Gray2Bin(c(1, 1, 1, 1))
}
\references{
Gray, Frank (1953):
Pulse Code Communication. US Patent 2 632 058.
}
