% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaGaDecode.R
\name{without}
\alias{without}
\title{Returns elements of
vector \code{x} without elements in \code{y}.}
\usage{
without(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
A vector.
}
\description{
Returns elements of
vector \code{x} without elements in \code{y}.
}
\examples{
a<-sample(1:15,15, replace=FALSE)
b<-c(1, 3, 5)
without(a, b)
}
\concept{Utility}
