% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaGaDecode.R
\name{xegaGaDecodeGene}
\alias{xegaGaDecodeGene}
\title{Decode a gene.}
\usage{
xegaGaDecodeGene(gene, lF)
}
\arguments{
\item{gene}{A binary gene (the genotype).}

\item{lF}{The local configuration of the genetic algorithm.}
}
\value{
The decoded gene (the phenotype).
}
\description{
\code{xegaGaDecodeGene()} decodes a binary gene.
}
\examples{
gene<-xegaGaInitGene(lFxegaGaGene)
xegaGaDecodeGene(gene, lFxegaGaGene)

}
\seealso{
Other Decoder: 
\code{\link{xegaGaGeneMap}()},
\code{\link{xegaGaGeneMapGray}()},
\code{\link{xegaGaGeneMapIdentity}()},
\code{\link{xegaGaGeneMapPerm}()}
}
\concept{Decoder}
