% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaGaInitGene.R
\name{xegaGaInitGene}
\alias{xegaGaInitGene}
\title{Generate a random binary gene.}
\usage{
xegaGaInitGene(lF)
}
\arguments{
\item{lF}{The local configuration of the genetic algorithm.}
}
\value{
A binary gene (a named list):
        \itemize{
        \item \code{$evaluated}: FALSE. See package \code{xegaSelectGene}.
        \item \code{$evalFail}:  FALSE. Set by the error handler(s)
                                 of the evaluation functions
                                 in package \code{xegaSelectGene} 
                                 in the case of failure.
        \item \code{$fit}:       Fitness.
        \item \code{$gene1}:     Binary gene.
        }
}
\description{
\code{xegaGaInitGene()} generates a random binary gene 
             with a given length.
}
\examples{
xegaGaInitGene(lFxegaGaGene)

}
\concept{Gene Generation.}
