% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaGaMutate.R
\name{xegaGaMutateGene}
\alias{xegaGaMutateGene}
\title{Mutate a gene.}
\usage{
xegaGaMutateGene(gene, lF)
}
\arguments{
\item{gene}{A binary gene.}

\item{lF}{The local configuration of the genetic algorithm.}
}
\value{
A binary gene.
}
\description{
\code{xegaGaMutateGene()} mutates a binary gene.
              The per-bit mutation rate is given by 
              \code{lF$BitMutationRate1()}.
}
\examples{
parm<-function(x) {function() {return(x)}}
lFxegaGaGene$BitMutationRate1<-parm(1.0)
gene1<-xegaGaInitGene(lFxegaGaGene)
xegaGaDecodeGene(gene1, lFxegaGaGene)
lFxegaGaGene$BitMutationRate1()
gene<-xegaGaMutateGene(gene1, lFxegaGaGene)
xegaGaDecodeGene(gene, lFxegaGaGene)
}
\seealso{
Other Mutation: 
\code{\link{xegaGaIVAdaptiveMutateGene}()}
}
\concept{Mutation}
